Refactor forcing. Default to base, rest to forcing.py
