+++ src/ewatercycle/plugins/hype/forcing.py
@@ -6,9 +6,11 @@ from typing import Literal, Optional
 import pandas as pd
 from esmvalcore.experimental import get_recipe
 
-from ewatercycle.forcing._default import DefaultForcing, _session
-from ewatercycle.forcing.datasets import DATASETS
+from ewatercycle.base.forcing import _session
+from ewatercycle.base.forcing import DATASETS
 from ewatercycle.util import get_time, to_absolute_path
+
+from ewatercycle.base.forcing import DefaultForcing
 
 
 class HypeForcing(DefaultForcing):
