+++ src/ewatercycle/plugins/wflow/forcing.py
@@ -3,9 +3,11 @@ from typing import Dict, Literal, Optional
 
 from esmvalcore.experimental import get_recipe
 
-from ewatercycle.forcing._default import DefaultForcing, _session
-from ewatercycle.forcing.datasets import DATASETS
+from ewatercycle.base.forcing import _session
+from ewatercycle.base.forcing import DATASETS
 from ewatercycle.util import get_extents, get_time, to_absolute_path
+
+from ewatercycle.base.forcing import DefaultForcing
 
 
 class WflowForcing(DefaultForcing):
